Create nqueens.py
From Coursera: Mathematical Thinking In Computer Science. On an n x n chessboard, can you place n queens such that they will not attack each other? Here is a python code that finds a permutation (array) that indicates in which row each queen should be (in the order of columns). With a board of 8x8, it returns a result almost immediately. With a board of 13 x 13 it takes more than 5 seconds.